+++ netbox/CHANGELOG.md
@@ -6,3 +6,4 @@
 * Fix formatting typos in README breaking hugo
 * Update docs for missing parameters
 * Use postgresql v16 because postgresql contrib is uninstalling v15 after installing
+* Update postgresql.conf file for new format with log_timezone and timezone